+++ src/lib/discovery/index.ts
@@ -3,7 +3,7 @@ import { shuffle } from "libp2p-gossipsub/src/utils";
 export { getPredefinedBootstrapNodes } from "./predefined";
 export * as predefined from "./predefined";
 export { Bootstrap, BootstrapOptions } from "./bootstrap";
-export { DnsClient, DnsNodeDiscovery, SearchContext } from "./dns";
+export * as dns from "./dns";
 export { Endpoints, DnsOverHttps } from "./dns_over_https";
 export { ENRTree, ENRTreeValues, ENRRootValues } from "./enrtree";